libmapi: Decoding GLOBSETs properly which have bitmask command
Two bugs were spotted:

* The *StartingValue* was set to a wrong data because the buffer_position
  was altered before stored additional data to combine in memory
* The GLOBSET ranges were incorrectly set because the calculated
  value was based on the bit value to check the mask against instead
  of the position in the loop.

Two unit tests were added which coverage the 100% of the
GLOBSET_parser_do_bitmask code.

The consequence of this bug is described as follows:

Once an idset is incorrectly decoded for one of the fxics meta-properties
when they are returned as the final result of an upload/download operation
in the FastTransferBuffer returns an inconsistent idset with incorrect
GLOBSET ranges, for instance, having a folder ids which it does not have
leading to the client to stop syncing that folder.

Conflicts:
	CHANGES.md